pyproject.toml
[project]
name = "ai-job-recommender-system-using-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apify-client>=1.12.0",
    "groq>=0.30.0",
    "langchain>=0.3.26",
    "langchain-groq>=0.3.6",
    "mcp[cli]>=1.11.0",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
]
